requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
transformers>=4.41.0
torch==2.0.1
numpy==1.24.3
pdfminer.six==20221105
python-docx==0.8.11
pydantic==1.10.8
python-multipart==0.0.5
safetensors>=0.4.1
beautifulsoup4==4.12.2
requests==2.31.0
google-generativeai==0.3.0
together==0.2.5
sentence-transformers==3.3.1
gunicorn==21.2.0
python-dotenv==1.0.0